Add expression text to tracebacks.
This is a bit hacky. (Inspired by a similar feature in `spy`, but
simpler.)

The actual traceback has an empty `text` for any code that comes from
a `<string>`, and there's really no way to distinguish the user's
input from some other string that gets evaluated farther down the
chain.

Also, if the expression does something like define a lambda and call
it (or pass it to another function that calls it later), it will also
be from file `<string>` with empty text. We could distinguish this
one, because the location will be, e.g., `<lambda>` instead of
`<module>`, but since we can't distinguish the previous case, there
doesn't seem to be much point.

So, what we do is just fill in the text for the first traceback line
we see from file `<string>`, and then not fill it in for later lines
from the same file.

Signed-off-by: Russell <russell.s.stewart@gmail.com>