Fix test suite
There were two problems with the test suite. The first one is that it lacked
the `__init__.py` file in the `test/` directory, so the tests couldn't be auto
discovered. The second one was that one of them was failing with a `TypeError:
Unicode-objects must be encoded before hashing` that can be easily fixed.